requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
croniter==6.0.0
discord-py-interactions==5.13.2
discord-typings==0.9.0
discord.py==2.4.0
emoji==2.14.1
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
mariadb==1.1.11
multidict==6.1.0
packaging==24.2
propcache==0.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
SQLAlchemy==2.0.37
tomli==2.2.1
typing_extensions==4.12.2
yarl==1.18.3
